services/auth-service/requirements.txt
flask
flask-sqlalchemy
python-dotenv
passlib
bcrypt==3.2.0
pymysql
cryptography
passlib
